more generic names: P2PTestNode -> OperaNodeStaff
